fix zip chinese file name error
